Moved tests to their own test folder, moved test store code into its own file, and cleaned up test file code
