+++ src/tests/App.test.js
@@ -1,21 +1,11 @@
 import React from 'react';
 import {debug, render, fireEvent, cleanup } from '@testing-library/react';
-import App from './components/Home';
+import App from '../components/Home';
 import {Provider, useSelector} from "react-redux";
 import { compose, applyMiddleware } from "redux";
-import logger from "redux-logger";
-import thunk from "redux-thunk";
 import { createMemoryHistory } from 'history'
 import { Router} from 'react-router-dom'
-
-import { createStore } from 'redux';
-import  rootReducer from './reducers';
-
-
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-
-const store = createStore(rootReducer, composeEnhancers(applyMiddleware(thunk, logger)));
-
+import {store} from "./store";
 
 test('App component renders to Homepage', () => {
   const history = createMemoryHistory()
