+++ src/tests/Home.test.js
@@ -1,23 +1,11 @@
 import React from 'react';
 import ReactTestUtils, {mount} from 'react-dom/test-utils';
 import {debug, render, fireEvent, cleanup} from '@testing-library/react';
-import Home from './components/Home';
-import Register from './components/Register';
+import Home from '../components/Home';
 import {Provider, useSelector} from "react-redux";
-import { compose, applyMiddleware } from "redux";
-import logger from "redux-logger";
-import thunk from "redux-thunk";
 import { createMemoryHistory } from 'history'
 import { Router} from 'react-router-dom'
-
-import { createStore } from 'redux';
-import  rootReducer from './reducers';
-
-
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-
-const store = createStore(rootReducer, composeEnhancers(applyMiddleware(thunk, logger)));
-
+import {store} from "./store";
 
 test('Home page renders links for register and login', async () => {
   const history = createMemoryHistory()
